Extending FLIP::parameters so that it contains exposure parameters as well, moving histogram and pooling out to a separate header (again).
